Fix Issue #91: Support for Gradle 4.x for Clover plugin
Migrated all code to work on Gradle 4.4.1. Solved the issues
with the new language specific classes directories by
completing the work to solve the joint Groovy/Java compile.

Fix Issue #55: Avoid Joint Compilation of Java Sources by Groovy Compiler

From now on the expectation is that code in groovy directories
will be compiled using the Groovy compiler for both Java and
Groovy code colocated. Java code in a java directory will
be compiled using the Java compiler. Previously when Groovy
was installed all code was compiled with Groovy without care.

Fix Issue #80: Use TestKit for defining and executing functional tests

Enabled the use of Gradle TestKit for functional testing.
This eliminated the dependency on arcane logic in the
build.gradle files that set plugin classpaths in weird ways.
We are now testing with multiple Gradle releases, currently
2.14.1, 3.5.1 and 4.4.1 are tested.

Fix Issue #96: Automate release process following gradle-docker-plugin scheme

The changes to include the refined Gradle build logic
from the Docker plugin were integrated.  We should have
a semi-automatic release process now based on tag pushing.

Other changes:

Removed the support for Clover license. Seems silly to keep
this logic when OpenClover is available. The convention
code is left to allow some level of compatibility.

Cleaned up the Java test code in all functional test data
projects to use Junit 4 annotations instead of Junit 3.

+++ src/main/groovy/com/bmuschko/gradle/clover/RestoreOriginalClassesAction.groovy
@@ -3,11 +3,13 @@ package com.bmuschko.gradle.clover
 import org.gradle.api.Action
 import org.gradle.api.Task
 
+import groovy.transform.CompileStatic
+
 class RestoreOriginalClassesAction implements Action<Task> {
-    Set<CloverSourceSet> sourceSets
-    Set<CloverSourceSet> testSourceSets
+    List<CloverSourceSet> sourceSets
+    List<CloverSourceSet> testSourceSets
 
-
+    @CompileStatic
     @Override
     void execute(Task t) {
         def ant = new AntBuilder()
@@ -18,9 +20,12 @@ class RestoreOriginalClassesAction implements Action<Task> {
         moveAllBackupDirsToClassesDirs(ant, getTestSourceSets())
     }
 
-    private void deleteAllClassesDirectories(AntBuilder ant, Set<CloverSourceSet> sourceSets) {
+    @CompileStatic
+    private void deleteAllClassesDirectories(AntBuilder ant, List<CloverSourceSet> sourceSets) {
         for(CloverSourceSet sourceSet : sourceSets) {
-            deleteClassesDirectory(ant, sourceSet.classesDir)
+            if (CloverSourceSetUtils.existsDirectory(sourceSet.classesDir)) {
+                deleteClassesDirectory(ant, sourceSet.classesDir)
+            }
          }
     }
 
@@ -30,14 +35,15 @@ class RestoreOriginalClassesAction implements Action<Task> {
         }
     }
 
-    private void moveAllBackupDirsToClassesDirs(AntBuilder ant, Set<CloverSourceSet> sourceSets) {
+    @CompileStatic
+    private void moveAllBackupDirsToClassesDirs(AntBuilder ant, List<CloverSourceSet> sourceSets) {
         for(CloverSourceSet sourceSet : sourceSets) {
             moveBackupToClassesDir(ant, sourceSet.backupDir, sourceSet.classesDir)
         }
     }
 
     private void moveBackupToClassesDir(AntBuilder ant, File backupDir, File classesDir) {
-        if(CloverSourceSetUtils.existsDirectory(backupDir)) {
+        if (CloverSourceSetUtils.existsDirectory(backupDir)) {
             ant.move(file: backupDir.canonicalPath, tofile: classesDir.canonicalPath, failonerror: true)
          }
     }
